Conflicts:
	Installer/VersionInfo.ini
	MultiMiner.CoinApi/Properties/AssemblyInfo.cs
	MultiMiner.CoinChoose/Properties/AssemblyInfo.cs
	MultiMiner.CoinWarz/Properties/AssemblyInfo.cs
	MultiMiner.Coinbase/Properties/AssemblyInfo.cs
	MultiMiner.Discovery/Properties/AssemblyInfo.cs
	MultiMiner.Engine/Properties/AssemblyInfo.cs
	MultiMiner.MobileMiner/Properties/AssemblyInfo.cs
	MultiMiner.Remoting/Properties/AssemblyInfo.cs
	MultiMiner.Services/Properties/AssemblyInfo.cs
	MultiMiner.Stats/Properties/AssemblyInfo.cs
	MultiMiner.Update/Properties/AssemblyInfo.cs
	MultiMiner.Utility/Properties/AssemblyInfo.cs
	MultiMiner.Win/Forms/MinerForm.cs
	MultiMiner.Win/Properties/AssemblyInfo.cs
	MultiMiner.Xgminer.Api/Properties/AssemblyInfo.cs
	MultiMiner.Xgminer.Discovery/Properties/AssemblyInfo.cs
	MultiMiner.Xgminer/Properties/AssemblyInfo.cs

+++ MultiMiner.Engine/Installers/MinerInstaller.cs
@@ -30,7 +30,10 @@ namespace MultiMiner.Engine.Installers
                 //first delete the folder contents. this became necessary with cgminer 3.8.0 because
                 //ck stopped shipping cgminer-nogpu.exe, which would leave an old executable behind
                 //and gum up the works later (running an older exe to find the installed version)
-                DeleteFolderContents(destinationFolder);
+
+                //only delete files and not folders though - we want to leave behind any OpenCL
+                //kernels the user may have installed
+                DeleteFolderContents(destinationFolder, true);
 
                 Unzipper.UnzipFileToFolder(minerDownloadFile, destinationFolder);
             }
@@ -40,7 +43,7 @@ namespace MultiMiner.Engine.Installers
             }
         }
 
-        private static void DeleteFolderContents(string folderPath)
+        private static void DeleteFolderContents(string folderPath, bool preserveKernels)
         {
             if (!Directory.Exists(folderPath))
                 //necessary or an Exception is thrown under Mono for OS X
@@ -49,12 +52,11 @@ namespace MultiMiner.Engine.Installers
             DirectoryInfo directoryInfo = new DirectoryInfo(folderPath);
 
             foreach (FileInfo fileInfo in directoryInfo.GetFiles())
-                fileInfo.Delete();
-
-            foreach (DirectoryInfo di in directoryInfo.GetDirectories())
             {
-                DeleteFolderContents(di.FullName);
-                di.Delete();
+                if (fileInfo.Extension.Equals(".cl", StringComparison.OrdinalIgnoreCase) && preserveKernels)
+                    continue;
+
+                fileInfo.Delete();
             }
         }
